+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ### Unreleased ###
 
+### 1.0.2: April 17th, 2018
+* Menu icons refactoring
+* Menu icon for Elementor
+* Menu icon for OceanWP
+* Menu icon for Hustle
+* Rename "Theme Panel" from Ocean to "Ocean"
+
 ### 1.0.1: April 10th, 2018
 * Gravity Forms menu icon
 * Scripts in footer